Adding favourites page, and home page,
All frontend for mobile done, We still need endpoints to get acces to db
