+++ client/src/components/Mini-cards/mini_card.module.css
@@ -1,0 +1,79 @@
+.mini_card_wrapper{
+    display: flex;
+    width: 10.0625rem;
+   
+    flex-direction: column;
+    align-items: flex-start;
+    flex-shrink: 0;
+
+    box-shadow: 0px 4px 20px 0px rgba(0, 0, 0, 0.10);
+
+    
+}
+.mini_card_img_wrapper{
+    flex: 1;
+    overflow: hidden;
+    position: relative;
+
+    border-radius: 1rem 1rem 0rem 0rem;
+    background: var(--grey);
+}
+.img_heart_wrap{
+    display: flex;
+    flex-direction: row;
+    width: 100%;
+    position: absolute;
+    justify-content: flex-end;
+    align-items: flex-start;
+    right: 4%;
+    top: 1%;
+}
+.img{
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+.mini_card_text{
+    display: flex;
+    flex-direction: column;
+    height: 50%;
+    width: 100%;
+}
+.mini_card_title{
+    display: flex;
+   flex-direction: row;
+   justify-content: flex-start;
+}
+.mini_card_title > p{
+    color: var(--default-font-clr);
+    /* Subtitle 2 */
+    font-family: var(--default-font);
+    font-size: 0.875rem;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 1.5rem; /* 171.429% */
+    letter-spacing: 0.00625rem;
+}
+.mini_card_sub_title{
+    display: flex;
+    flex-direction: row;
+    gap: .5rem;
+    width: 100%;
+    justify-content: flex-start;
+    align-items: center;
+}
+.mini_card_sub_title > p{
+    color: var(--default-font-clr);
+    /* Caption */
+    font-family: var(--default-font);
+    font-size: 0.75rem;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 1rem; /* 133.333% */
+    letter-spacing: 0.025rem;
+}
+.mini_subText{
+    display: flex;
+    padding-left: .5rem;
+    flex-direction: column;
+}
