+++ client/src/components/Personal_preview/personal_preview.module.css
@@ -18,105 +18,7 @@
     line-height: 1rem; /* 133.333% */
     letter-spacing: 0.025rem;
 }
-.img_container{
-    height: 11.5rem;
-    align-self: stretch;
-    border-radius: 1rem 1rem 0rem 0rem;
-    background: var(--grey);
-}
-.img_display_area{
 
-}
-
-.menu_btns{
-    display: flex;
-    padding: 0.5rem 1rem;
-    align-items: center;   
-    justify-content: center;
-}
-.menu_btns > button{
-    background: none;
-
-    font-family: var(--default-font);
-    color: var(--Blue-40);
-    font-size: 1rem;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 1.5rem; 
-    letter-spacing: 0.00938rem;
-    border: none;
-    position: relative;
-}
-.menu_btns > button::after {
-    content: '';
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 100%;
-    height: 2px;
-    background-color: var(--Blue-40); /* Set the initial color to grey */
-    transition: background-color 0.3s ease; /* Add transition for smooth color change */
-  }
-  
-  .menu_btns > button.selected::after {
-    background-color: red; /* Change to red for the selected category */
-  }
-  .selected{
-    color: var(--default-font-clr)!important; 
-}
-
-.redBox{
-    border-radius: 1rem;
-    border: 1px solid var(--secondary-BG);
-    background: var(--default-BG);
-    /* Box Shadow Jen */
-    box-shadow: 0px 4px 20px 0px rgba(0, 0, 0, 0.10);
-    margin-bottom: 2rem;
-    display: flex;
-    width: 21.375rem;
-    padding:  1rem 1rem 2rem 1rem;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 0.5rem;
-}
-
-.user_account{
-
-}
-.user_details{
-    display: flex;
-    flex-direction: row;
-    gap: .5rem;
-}
-.user_name_wrapper{
-
-}
-.user_header{
-    color: var(--default-font-clr);
-    /* H5 */
-    font-family: var(--default-font);
-    font-size: 1.5rem;
-    font-style: normal;
-    font-weight: 700;
-    line-height: 1.5rem; /* 100% */
-    letter-spacing: 0.01125rem;
-}
-.role_details{
-    display: flex;
-    flex-direction: row;
-    gap: 2.5rem;
-}
-.user_role{
-
-}
-.details_submenu{
-    width: 100%;
-}
-
-.usher{
-display: flex;
-flex-direction: row;
-}
 .lusher{
 
 }
