+++ client/src/views/Favourites/Favourites.jsx
@@ -1,36 +1,36 @@
 import { useState } from 'react';
 import { Nav } from '../index.js';
-import { briefcase, locationBlack, search, sliders, userSml } from '../../assets/Icons/index.js';
-import style from './home.module.css';
-import { heartlight } from '../../assets/Icons/dropdownicons/index.js';
 import { Mini_card, Spacer_bottom } from '../../components/index.js';
+import { briefcase, locationBlack, search, sliders, userSml } from '../../assets/Icons/index.js';
+import style from './favo.module.css';
 
-function Home(){
+function Favourites(){
     const [showFilter, setShowFilter] = useState(false);
     const [animationReverted, setAnimationReverted] = useState(false);
 
-    /*Dummy data */
-    sessionStorage.setItem('userRole', 'company');
-    sessionStorage.setItem('companyName', ' Big company');
-   
-    sessionStorage.setItem('firstname', 'Hannes');
-    sessionStorage.setItem('lastName', 'Hansson');
+    const handleFilter = () => {
+        setShowFilter(!showFilter);
+        setAnimationReverted(false); // Reset animation reverted state
+    }
     
+    const handleAnimationEnd = () => {
+        setAnimationReverted(true); // Set animation reverted state after animation ends
+    }
 
-    const userRole = 'company';
-    const compName = sessionStorage.getItem('companyName');
-    const firstName = sessionStorage.getItem('firstname');
-    const lastName = sessionStorage.getItem('lastName');
-/*Dummy data */
-
-const handleFilter = () => {
-    setShowFilter(!showFilter);
-    setAnimationReverted(false); // Reset animation reverted state
-}
-
-const handleAnimationEnd = () => {
-    setAnimationReverted(true); // Set animation reverted state after animation ends
-}
+     /*Dummy data */
+     sessionStorage.setItem('userRole', 'company');
+     sessionStorage.setItem('companyName', ' Big company');
+    
+     sessionStorage.setItem('firstname', 'Hannes');
+     sessionStorage.setItem('lastName', 'Hansson');
+     
+ 
+     const userRole = 'company';
+     const compName = sessionStorage.getItem('companyName');
+     const firstName = sessionStorage.getItem('firstname');
+     const lastName = sessionStorage.getItem('lastName');
+ /*Dummy data */
+ 
     return(
         <>
             <div className={style.main}>
@@ -73,53 +73,10 @@ const handleAnimationEnd = () => {
                     </div>
                 </div>
 
-                <Spacer_bottom />
-
-                <div className={style.quiz_wrapper}>
-                    <button className={style.quizbtn}>
-                        <p className={style.quiz_head}>What company suits you the best?</p>
-                        <p className={style.quiz_sub}>Increase your chances of connecting by taking this quiz</p>
-                    </button>
-                </div>
-
-                {/*SLIDER GOES HERE */}
-                <div className={style.new_companies_slider}>
-
-                    <div className={style.new_companies_slide_card}>
-                        <div className={style.new_cmp_head}>
-                            <p>New companies</p>
-                        </div>
-                        <div className={style.redBox}>
-                            <div className={style.img_wrapper}>
-                                <div className={style.img_display_area}>
-                                    <div className={style.icon_container}>
-                                        <img src={heartlight} />
-                                    </div>
-                                </div>
-                            </div>
-
-                            <div className={style.name_box}>
-                                <img src={briefcase} />
-                                <p>{compName}</p>
-                            </div>
-                            <div className={style.name_box}>
-                                <div className={style.name_loc_wrapper}>
-                                    <img src={userSml} />
-                                    <p>{firstName} {lastName}</p>
-                                </div>
-                                <div className={style.name_loc_wrapper}>                                
-                                    <img src={locationBlack} />
-                                    <p>Company City</p>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-
-                </div>
-                {/*END SLIDER */}
+               
 
                 <div className={style.attending_comps}>
-                    <p>All companies attending</p>
+                    <p>All Favourites</p>
 
                     <div className={style.mini_cards_containter}>
                         {/* FOREACH COMPANY in DB  ADD A MINI CARD */}
@@ -148,9 +105,10 @@ const handleAnimationEnd = () => {
                         />
                     </div>
                 </div>
+
             </div>
         </>
     );
 }
 
-export default Home
+export default Favourites
