+++ client/src/views/Favourites/favo.module.css
@@ -135,134 +135,8 @@
     letter-spacing: 0.07813rem;
     position: relative;
   }
-  
-  .quiz_wrapper {}
-  
-  .quizbtn {
-    display: flex;
-    width: 21.375rem;
-    padding: 1.5rem 1rem;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 0.5rem;
-    border-radius: 1rem;
-    background: var(--secondary-BG);
-    /* Box Shadow Jen */
-    box-shadow: 0px 4px 20px 0px rgba(0, 0, 0, 0.10);
-    border: 1px solid var(--secondary-BG);
-  }
-  
-  .quiz_head {
-    align-self: stretch;
-    color: var(--secondary-font-clr);
-    /* Subtitle 1 */
-    font-family: var(--default-font);
-    font-size: 1rem;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 1.5rem; /* 150% */
-    letter-spacing: 0.00938rem;
-  }
-  
-  .quiz_sub {
-    color: var(--secondary-font-clr);
-    /* Caption */
-    font-family: var(--default-font);
-    font-size: 0.75rem;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 1rem; /* 133.333% */
-    letter-spacing: 0.025rem;
-  }
-  
-  .quizbtn > p {
-    margin: 0;
-  }
-  
-  .new_companies_slide_card {
-    display: flex;
-    flex-direction: column;
-  }
-  
-  .new_cmp_head {}
-  
-  .new_cmp_head > p {
-    color: var(--Blue-80);
-    /* Subtitle 2 */
-    font-family: var(--default-font);
-    font-size: 0.875rem;
-    font-style: normal;
-    font-weight: 400;
-    line-height: 1.5rem; /* 171.429% */
-    letter-spacing: 0.00625rem;
-  }
-  
-  .redBox {
-    display: flex;
-    /*  width: 21.375rem; */
-  
-    padding: 2rem 1.5rem;
-  
-    flex-direction: column;
-    justify-content: center;
-    align-items: flex-start;
-    gap: 0.5rem;
-  
-    border-radius: 1rem;
-    border: 1px solid var(--secondary-BG);
-    background: var(--default-BG);
-    /* Box Shadow Jen */
-    box-shadow: 0px 4px 20px 0px rgba(0, 0, 0, 0.10);
-    margin-bottom: 2rem;
-  }
-  
-  .img_wrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-  }
-  
-  .img_display_area {
-    height: 11.5rem;
-    width: 100%;
-    align-self: stretch;
-    border-radius: 1rem 1rem 0rem 0rem;
-    background: var(--grey);
-  }
-  
-  .icon_container {
-    display: flex;
-    flex-direction: row;
-    width: 100%;
-    align-items: flex-start;
-    justify-content: flex-end;
-  }
-  
-  .name_box {
-    display: flex;
-    flex-direction: row;
-    gap: .5rem;
-  }
-  
-  .name_box > p {
-    color: var(--default-font-clr);
-    /* Subtitle 2 */
-    font-family: var(--default-font);
-    font-size: 1.5rem;
-    font-style: normal;
-    font-weight: 700;
-    line-height: 1.5rem; /* 100% */
-    letter-spacing: 0.01125rem;
-  }
-  
-  .name_loc_wrapper {
-    display: flex;
-    flex-direction: row;
-    gap: .5rem;
-  }
-  
+
+
   .attending_comps {
     display: flex;
     width: 100%;
@@ -271,7 +145,7 @@
     align-items: flex-start;
     gap: 1rem;
   }
-  
+
   .attending_comps > p {
     color: var(--Blue-80);
     /* Subtitle 2 */
